pyproject.toml
[tool.poetry]
name = "discord-mirror-listener"
version = "0.1.0"
description = ""
authors = ["Jonathan Miller <jon@jmill.dev>"]
packages = [{ include = "listener" }]

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^0.18.0"
"discord.py" = "^1.7.3"
cogwatch = "^2.1.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.2"
black = "^21.6b0"
mypy = "^0.910"
pytest = "^6.2.4"
ipykernel = "^6.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
listener = 'listener.bot:start'
